Refactor vedicts into experiment_results and blocking_events
